Correções acessos

+++ Hydra.Such.Portal/Hydra.Such.Portal/Controllers/AdministracaoController.cs
@@ -2760,47 +2760,47 @@ namespace Hydra.Such.Portal.Controllers
             UserAccessesViewModel UPerm = new UserAccessesViewModel();
             if (id == "Engenharia")
             {
-                UPerm = DBUserAccesses.GetByUserAreaFunctionality(User.Identity.Name, (int)Enumerations.Areas.Engenharia, (int)Enumerations.Features.Administração);
+                UPerm = DBUserAccesses.GetByUserAreaFunctionality(User.Identity.Name, Enumerations.Areas.Engenharia, Enumerations.Features.Administração);
             }
             if (id == "Ambiente")
             {
-                UPerm = DBUserAccesses.GetByUserAreaFunctionality(User.Identity.Name, (int)Enumerations.Areas.Ambiente, (int)Enumerations.Features.Administração);
+                UPerm = DBUserAccesses.GetByUserAreaFunctionality(User.Identity.Name, Enumerations.Areas.Ambiente, Enumerations.Features.Administração);
             }
             if (id == "Nutricao")
             {
-                UPerm = DBUserAccesses.GetByUserAreaFunctionality(User.Identity.Name, (int)Enumerations.Areas.Nutrição, (int)Enumerations.Features.Administração);
+                UPerm = DBUserAccesses.GetByUserAreaFunctionality(User.Identity.Name, Enumerations.Areas.Nutrição, Enumerations.Features.Administração);
             }
             if (id == "Vendas")
             {
-                UPerm = DBUserAccesses.GetByUserAreaFunctionality(User.Identity.Name, (int)Enumerations.Areas.Vendas, (int)Enumerations.Features.Administração);
+                UPerm = DBUserAccesses.GetByUserAreaFunctionality(User.Identity.Name, Enumerations.Areas.Vendas, Enumerations.Features.Administração);
             }
             if (id == "Apoio")
             {
-                UPerm = DBUserAccesses.GetByUserAreaFunctionality(User.Identity.Name, (int)Enumerations.Areas.Apoio, (int)Enumerations.Features.Administração);
+                UPerm = DBUserAccesses.GetByUserAreaFunctionality(User.Identity.Name, Enumerations.Areas.Apoio, Enumerations.Features.Administração);
             }
             if (id == "PO")
             {
-                UPerm = DBUserAccesses.GetByUserAreaFunctionality(User.Identity.Name, (int)Enumerations.Areas.PO, (int)Enumerations.Features.Administração);
+                UPerm = DBUserAccesses.GetByUserAreaFunctionality(User.Identity.Name, Enumerations.Areas.PO, Enumerations.Features.Administração);
             }
             if (id == "NovasAreas")
             {
-                UPerm = DBUserAccesses.GetByUserAreaFunctionality(User.Identity.Name, (int)Enumerations.Areas.NovasÁreas, (int)Enumerations.Features.Administração);
+                UPerm = DBUserAccesses.GetByUserAreaFunctionality(User.Identity.Name, Enumerations.Areas.NovasÁreas, Enumerations.Features.Administração);
             }
             if (id == "Internacionalizacao")
             {
-                UPerm = DBUserAccesses.GetByUserAreaFunctionality(User.Identity.Name, (int)Enumerations.Areas.Internacional, (int)Enumerations.Features.Administração);
+                UPerm = DBUserAccesses.GetByUserAreaFunctionality(User.Identity.Name, Enumerations.Areas.Internacional, Enumerations.Features.Administração);
             }
             if (id == "Juridico")
             {
-                UPerm = DBUserAccesses.GetByUserAreaFunctionality(User.Identity.Name, (int)Enumerations.Areas.Jurídico, (int)Enumerations.Features.Administração);
+                UPerm = DBUserAccesses.GetByUserAreaFunctionality(User.Identity.Name, Enumerations.Areas.Jurídico, Enumerations.Features.Administração);
             }
             if (id == "Compras")
             {
-                UPerm = DBUserAccesses.GetByUserAreaFunctionality(User.Identity.Name, (int)Enumerations.Areas.Compras, (int)Enumerations.Features.Administração);
+                UPerm = DBUserAccesses.GetByUserAreaFunctionality(User.Identity.Name, Enumerations.Areas.Compras, Enumerations.Features.Administração);
             }
             if (id == "Administracao")
             {
-                UPerm = DBUserAccesses.GetByUserAreaFunctionality(User.Identity.Name, (int)Enumerations.Areas.Administração, (int)Enumerations.Features.Administração);
+                UPerm = DBUserAccesses.GetByUserAreaFunctionality(User.Identity.Name, Enumerations.Areas.Administração, Enumerations.Features.Administração);
             }
 
             return UPerm;
